Sensors now record their minimum and maximum values per session
This is temporarily being displayed in the same string as the current
value, as I need to get some data out of my car tomorrow, but when I
refactor the GUI I will be making this display much neater.

Also pushed the update interval to 250ms to see how OBD2 copes with it.