Changed GUI to have sensors on one side and additional info on the other. + Turbo Timer!
The GUI is split into two vertical columns now. The left column has the
usual sensor display and the right column has a text box which shows
information from other features, such as the new turbo timer.
The turbo timer is currently set to a 120 second cooldown time. This may
be modified in future to increase the long the engine is running for.
You can also now fetch sensors by their shortname, which allows sensor
and feature objects to pull values from other sensors.

Some of the code does need refactoring, such as the use of magic
numbers, but at the moment this is a quick working version for me to use
whilst I'm out.